修改 jest测试样例
