+++ .eslintrc.js
@@ -4,6 +4,7 @@ module.exports = {
     node: true
   },
   extends: ["plugin:vue/essential", "@vue/prettier", "@vue/typescript"],
+  //extends: ["@vue/prettier", "@vue/typescript"],
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
