+++ jest.config.js
@@ -3,12 +3,15 @@ module.exports = {
     'js',
     'jsx',
     'json',
-    'vue'
+    'vue',
+    'ts',
+    'tsx'
   ],
   transform: {
     '^.+\\.vue$': 'vue-jest',
     '.+\\.(css|styl|less|sass|scss|png|jpg|ttf|woff|woff2)$': 'jest-transform-stub',
-    '^.+\\.jsx?$': 'babel-jest'
+    '^.+\\.jsx?$': 'babel-jest',
+    "^.+\\.tsx?$": "ts-jest"
   },
   moduleNameMapper: {
     '^@/(.*)$': '<rootDir>/src/$1'
@@ -19,5 +22,10 @@ module.exports = {
   testMatch: [
     '**/tests/unit/**/*.spec.(js|jsx|ts|tsx)|**/__tests__/*.(js|jsx|ts|tsx)'
   ],
-  testURL: 'http://localhost/'
+  testURL: 'http://localhost/',
+  globals: {
+    "ts-jest": {
+      babelConfig: true
+    }
+  }
 }
