+++ package.json
@@ -35,11 +35,11 @@
     "vue-property-decorator": "^8.2.1",
     "vue-router": "^3.0.3",
     "vuex": "^3.0.1",
-    "vuex-class": "^0.3.2",
     "vuex-module-decorators": "^0.10.1"
   },
   "devDependencies": {
     "@kazupon/vue-i18n-loader": "^0.4.1",
+    "@types/jest": "^24.0.18",
     "@types/js-cookie": "^2.2.2",
     "@vue/cli-plugin-babel": "^3.11.0",
     "@vue/cli-plugin-eslint": "^3.11.0",
@@ -52,13 +52,14 @@
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
     "babel-jest": "^23.6.0",
-    "eslint": "^6.3.0",
+    "eslint": "^5.16.0",
     "eslint-plugin-prettier": "^3.1.0",
     "eslint-plugin-vue": "^5.0.0",
     "prettier": "^1.18.2",
     "sass": "^1.18.0",
     "sass-loader": "^7.1.0",
     "svg-sprite-loader": "^4.1.3",
+    "ts-jest": "^23.0.0",
     "typescript": "^3.4.3",
     "vue-template-compiler": "^2.6.10"
   }
