--- src/views/demo/page1/index.vue
@@ -1,25 +1,0 @@
-<template>
-  <d2-container>
-    <template slot="header">Page 1 header</template>
-    Hello World
-  </d2-container>
-</template>
-
-<script lang="ts">
-import Component from 'vue-class-component'
-import { Vue } from 'vue-property-decorator'
-import { getModule } from 'vuex-module-decorators'
-import store from '../../../store';
-
-
-
-@Component({
-  name: 'page1'
-})
-export default class page1 extends Vue {
-
-  name: 'page1'
-  async mounted() {
-  }
-}
-</script>
